+++ zarrs/src/array/chunk_key_encoding/v2.rs
@@ -83,7 +83,8 @@ impl ChunkKeyEncodingTraits for V2ChunkKeyEncoding {
         let configuration = V2ChunkKeyEncodingConfiguration {
             separator: self.separator,
         };
-        MetadataV3::new_with_serializable_configuration(IDENTIFIER, &configuration).unwrap()
+        MetadataV3::new_with_serializable_configuration(IDENTIFIER.to_string(), &configuration)
+            .unwrap()
     }
 
     fn encode(&self, chunk_grid_indices: &[u64]) -> StoreKey {